% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot.R
\name{add_pepfar_rollup}
\alias{add_pepfar_rollup}
\title{Add PEPFAR Rollup}
\usage{
add_pepfar_rollup(df, sel_cntry)
}
\arguments{
\item{df}{UNAIDS data frame loaded from \code{load_unaids()}}

\item{sel_cntry}{PEPFAR country to visualize ("ALL PEPFAR" as default or list OU names)}
}
\description{
If desired by user, add in a PEPFAR roll up
}
\keyword{internal}
