% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{apply_indicator_type}
\alias{apply_indicator_type}
\title{Apply Indicator Type}
\usage{
apply_indicator_type(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Specify if the indicator's value is a percent, integer, or rate
}
\keyword{internal}
