% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{base_gray}
\alias{base_gray}
\title{Base gray for plot elements}
\format{
An object of class \code{character} of length 1.
}
\usage{
base_gray
}
\description{
color used for plot elements
}
\keyword{datasets}
