% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plot.R
\name{base_plot}
\alias{base_plot}
\title{95's Table Plot}
\usage{
base_plot(
  df,
  cntry,
  denom = c("PLHIV", "Relative"),
  grp = c("All", "Female 15+", "Male 15+", "Peds <15"),
  yr = NULL
)
}
\arguments{
\item{df}{dataframe from \code{load_unaids}}

\item{cntry}{PEPFAR country to visualize (list OU name)}

\item{grp}{age/sex group, c("All", "Female 15+", "Male 15+", "Peds <15")}

\item{yr}{year in question, defaults to max year}

\item{base}{Returns one of 2 table types eg "PLHIV" (default), "Relative"}
}
\value{
gt table
}
\description{
This function creates a summary table showing OU progress toward the 95-95-95's
}
\examples{
 \dontrun{
  df_unaids <- load_unaids()
   base_plot(df_unaids, "Kenya", denom = "PLHIV")
   base_plot(df_unaids, "Kenya", denom = "Relative")
}

}
\seealso{
\code{\link[=load_unaids]{load_unaids()}}
}
