% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{blue_palette}
\alias{blue_palette}
\title{Blue palette}
\format{
An object of class \code{character} of length 3.
}
\usage{
blue_palette
}
\description{
color used for dots/lines/bars
}
\keyword{datasets}
