% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{calc_95_goals}
\alias{calc_95_goals}
\title{Calculate the 95 targets for relative/plhiv base}
\usage{
calc_95_goals(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Calculate the 95 targets for relative/plhiv base
}
\keyword{internal}
