% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connected_scatterplot.R
\name{calculate_axis_limits}
\alias{calculate_axis_limits}
\title{Calculate Axis Limits}
\usage{
calculate_axis_limits(filtered_data)
}
\arguments{
\item{filtered_data}{A data frame containing columns \code{deaths} and \code{infections} to be used for axis limit calculation.}
}
\value{
A numeric vector of length two containing the global minimum and maximum values.
}
\description{
This internal function calculates the global minimum and maximum values for the x and y axes.
}
\keyword{internal}
