% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{caption_gray}
\alias{caption_gray}
\title{Text gray for plot caption}
\format{
An object of class \code{character} of length 1.
}
\usage{
caption_gray
}
\description{
color used for plot caption
}
\keyword{datasets}
