% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{clean_cols}
\alias{clean_cols}
\title{Clean Columns
Reorder and rename columns}
\usage{
clean_cols(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Clean Columns
Reorder and rename columns
}
\keyword{internal}
