% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{combo_plot}
\alias{combo_plot}
\title{Combo Plot}
\usage{
combo_plot(ctry_sel, folderpath_output = NULL)
}
\arguments{
\item{ctry_sel}{country to visualize}

\item{folderpath_output}{}
}
\value{
combined plots of 90-90-90 trends and ART gap by age/sex
}
\description{
Combo Plot
}
