% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connected_scatterplot.R
\name{construct_epi_plot}
\alias{construct_epi_plot}
\title{Construct Epidemic Control Plot}
\usage{
construct_epi_plot(filtered_data, min_max_year, axis_limits, year_limits)
}
\arguments{
\item{filtered_data}{A data frame containing the filtered data with columns \code{deaths}, \code{infections}, and \code{country}.}

\item{min_max_year}{A data frame containing the data points for the minimum and maximum years for each country.}

\item{axis_limits}{A numeric vector of length two containing the global minimum and maximum values for the axes.}

\item{year_limits}{A numeric vector of length two containing the global min and max of the year ranges}
}
\value{
A ggplot2 plot object that visualizes the epidemic control trajectories.
}
\description{
This internal function builds a ggplot object to visualize the relationship between deaths and infections.
}
\keyword{internal}
