% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{convert_numeric_string}
\alias{convert_numeric_string}
\title{Convert Numeric Values stored as string}
\usage{
convert_numeric_string(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Convert Numeric Values stored as string
}
\keyword{internal}
