% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{flag_95s}
\alias{flag_95s}
\title{Flag 95 Achievement}
\usage{
flag_95s(df, epi_95s_flag = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{epi_95s_flag}{add variables to dataframe for UNAIDS 95s achievement and
epidemic control status, default = TRUE}
}
\description{
Add flags for achievement across all three 95s (for both relative and PLHIV
base) across country, year, age, and sex.
}
\keyword{internal}
