% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{flag_epi}
\alias{flag_epi}
\title{Flag Epi contol status}
\usage{
flag_epi(df, epi_95s_flag = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{epi_95s_flag}{add variables to dataframe for UNAIDS 95s achievement and
epidemic control status, default = TRUE}
}
\description{
Add flags for Epi contol status for all age and sex
}
\keyword{internal}
