% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{grid_gray}
\alias{grid_gray}
\title{Grid gray for plot}
\format{
An object of class \code{character} of length 1.
}
\usage{
grid_gray
}
\description{
color used for plot grid
}
\keyword{datasets}
