% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_id.R
\docType{data}
\name{gs_clean_id_2022}
\alias{gs_clean_id_2022}
\title{2022 Clean Estimates - all data}
\format{
An object of class \code{character} of length 1.
}
\usage{
gs_clean_id_2022
}
\value{
Dataframe of all tidied UNAIDS 2022 Data
}
\description{
2022 Clean Estimates - all data
}
\keyword{internal}
