% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_id.R
\docType{data}
\name{gs_id_names}
\alias{gs_id_names}
\title{UNAIDS crosswalk ID}
\format{
An object of class \code{character} of length 1.
}
\usage{
gs_id_names
}
\value{
UNAIDS crosswalk ID
}
\description{
UNAIDS crosswalk ID
}
\keyword{internal}
