% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_unaids.R
\name{handle_historic}
\alias{handle_historic}
\title{Handle the import of historic data}
\usage{
handle_historic(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
The structure adjusted in v2.0 \link{2024-10-01}. This function provides the
handling of column names that have changed
}
\keyword{internal}
