% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_map}
\alias{indicator_map}
\title{Mapping of EDMS Indicator Names to those used in OHA}
\format{
A data frame with 21 rows and 3 columns:
\describe{
\item{acronym}{EDMS Acronym}
\item{indicator_edms}{Parsed EDMS indicator name}
\item{indicator}{Common indicator name}
}
}
\usage{
indicator_map
}
\description{
A mapping of the EDMS \code{acronym} to common names used by OHA analysts.
}
\keyword{datasets}
