% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_validation}
\alias{indicator_validation}
\title{Full list of expected indicators}
\format{
A data frame with 62 rows and 4 columns:
\describe{
\item{indicator_edms}{Parsed EDMS indicator name}
\item{age, sex}{EDMS indicator disaggregates}
\item{source}{Source folder on EDMS}
}
}
\usage{
indicator_validation
}
\description{
A dataframe containing the different indicators and their disaggregate
components to ensure all data expected for are being exported from EDMS.
}
\keyword{datasets}
