% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data for viz}
\usage{
load_data(usaid_email = NULL)
}
\arguments{
\item{usaid_email}{USAID email to store file on GDrive}
}
\description{
Load data for viz
}
\examples{
\dontrun{
   load_data("twhitmore@usaid.gov")}
}
