% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_unaids.R
\name{load_unaids}
\alias{load_unaids}
\title{Load tidy UNAIDS HIV Estimates and Test & Treat Data}
\usage{
load_unaids(pepfar_only = TRUE)
}
\arguments{
\item{pepfar_only}{filters dataset to only PEPFAR countries if TRUE (default = TRUE)}
}
\value{
df
}
\description{
This function is used to load the latest UNAIDS data. These data were queried
and exported from \href{https://edms.unaids.org/}{UNAIDS EMDS database}, cleaned
using \code{munge_edms} and finally loaded up on GitHub as a release. This
function downloads the data and loads it into your working session.
}
\examples{
 \dontrun{
   df_unaids <- load_unaids(pepfar_only = TRUE)
}

}
