% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{map_indicator}
\alias{map_indicator}
\title{Map working indicator name to EDMS indicator name}
\usage{
map_indicator(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Map working indicator name to EDMS indicator name
}
\keyword{internal}
