% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindthegap-package.R
\docType{package}
\name{mindthegap-package}
\alias{mindthegap}
\alias{mindthegap-package}
\title{mindthegap: Mind the Gap: 90-90-90 Trends and ART Coverage Gaps}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Package to pull UNAIDS estimates as well as plot trends in UNAIDS 90-90-90 goals and ART coverage gap by country.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USAID-OHA-SI/mindthegap}
  \item \url{https://usaid-oha-si.github.io/mindthegap}
  \item \url{https://usaid-oha-si.github.io/mindthegap/}
  \item Report bugs at \url{https://github.com/USAID-OHA-SI/mindthegap/issues}
}

}
\author{
\strong{Maintainer}: Karishma Srikanth \email{ksrikanth@usaid.gov}

Authors:
\itemize{
  \item Aaron Chafetz \email{achafetz@usaid.ogv}
  \item Lemlem Baraki \email{lbaraki@usaid.gov}
  \item Tim Essam \email{tessam@usaid.gov}
}

}
\keyword{internal}
