% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{munge_components}
\alias{munge_components}
\title{Clean up indicator and disaggregate columns}
\usage{
munge_components(df)
}
\arguments{
\item{df}{datafrane}
}
\description{
Clean up indicator and disaggregate columns
}
\keyword{internal}
