% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{munge_country}
\alias{munge_country}
\title{Munge PEPFAR country}
\usage{
munge_country(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Apply PEPFAR naming conventions where applicable and flag the countries that
received PEPFAR funding.
}
\keyword{internal}
