% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{munge_edms}
\alias{munge_edms}
\title{Clean UNAIDS Data from EDMS}
\usage{
munge_edms(path)
}
\arguments{
\item{path}{fileoath to the EDMS export (csv)}
}
\value{
df
}
\description{
Clean UNAIDS Data from EDMS
}
\examples{
 \dontrun{
  filepath <- "../DataList_10_1_2030-12_00_00-AM.csv"
  df <- munge_edms(filepath)
}
}
