% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_unaids.R
\name{munge_unaids}
\alias{munge_unaids}
\title{Cleaning UNAIDS Data}
\usage{
munge_unaids(return_type)
}
\arguments{
\item{return_type}{type of Data to return}
}
\value{

}
\description{
This function cleans and fetches UNAIDS Estimates/Test and Treat Data
}
\examples{
 \dontrun{
   munge_unaids(sheet_id = googledrive::as_id('1tkwP532mPL_yy7hJuHNAHaZ1_K_wd7zo_8AjeOe7fRs'), tab = 3, skipnum = 4)
}

}
