% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{parse_indicator}
\alias{parse_indicator}
\title{Parse indicator
Standardize indicator across years and other components (eg bounds)}
\usage{
parse_indicator(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Parse indicator
Standardize indicator across years and other components (eg bounds)
}
\keyword{internal}
