% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pepfar}
\alias{pepfar}
\title{PEPFAR countries + ISO codes}
\format{
A data frame with 55 rows and 2 columns:
\describe{
\item{country_pepfar}{Country name}
\item{iso3}{Country ISO3 code}
}
}
\usage{
pepfar
}
\description{
This is a dataframe of countries that receive PEPFAR funding. This df comes
from \code{glamr::pepfar_country_list} and is originally derived from the
PEPFAR Financial Structured Dataset (FSD).
}
\keyword{datasets}
