% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_id.R
\docType{data}
\name{pepfar_clean_id_2022}
\alias{pepfar_clean_id_2022}
\title{2022 Clean Estimates - PEPFAR only data}
\format{
An object of class \code{character} of length 1.
}
\usage{
pepfar_clean_id_2022
}
\value{
Dataframe of PEPFAR only tidy UNAIDS 2022 Data
}
\description{
2022 Clean Estimates - PEPFAR only data
}
\keyword{internal}
