% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connected_scatterplot.R
\name{plot_connected_scatter}
\alias{plot_connected_scatter}
\title{Plot Epidemic Control Trajectories}
\usage{
plot_connected_scatter(.data, sel_cntry, year_range = 2010:2023)
}
\arguments{
\item{.data}{A data frame containing epidemiological data. It should include columns such as \code{country}, \code{year}, \code{indicator}, \code{estimate}, and others required for processing.}

\item{sel_cntry}{A character vector of countries to include in the plot.}

\item{year_range}{A numeric vector specifying the range of years to include in the analysis. Default is \code{2010:2023}.}
}
\value{
A ggplot2 plot object showing the epidemic control trajectories with deaths on the x-axis and new infections on the y-axis.
}
\description{
Generates a connected scatterplot of trajectories for deaths and new infections for specified countries.
}
\details{
This function serves as a wrapper, orchestrating the preparation, calculation, and visualization of the data. It processes the input data, calculates appropriate axis limits, constructs the epidemic control plot, and finally adds an annotation for epidemic control.
}
\examples{
 \dontrun{
   df_unaids <- load_unaids(pepfar_only = FALSE)
   plot_connected_scatter(df_unaids)
   plot_connected_scatter(df_unaids, sel_cntry = "Lesotho")
   plot_connected_scatter(df_unaids, sel_cntry = c("South Africa", "Zambia", "Kenya", "Malawi"))
}
}
