% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connected_scatterplot.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare Data for Plotting}
\usage{
prepare_plot_data(df_epi_fltrd)
}
\arguments{
\item{df_epi_fltrd}{A data frame created by \code{prep_epi_data} containing epidemiological data filtered by country and year.}
}
\value{
A list containing the original filtered data (\code{filtered_data}) and a data frame with the minimum and maximum year data (\code{min_max_year}).
}
\description{
This internal function identifies the minimum and maximum year data points for use in plotting geometries. It could
probably be removed but is kept in case the \code{df_epi_fltrd} data frame is large.
}
\keyword{internal}
