% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot.R
\name{prepare_viz_data}
\alias{prepare_viz_data}
\title{Prepare viz data}
\usage{
prepare_viz_data(df, df_edms, sel_cntry)
}
\arguments{
\item{df}{UNAIDS data frame loaded from \code{load_unaids()}}

\item{df_edms}{original df provided}

\item{sel_cntry}{PEPFAR country to visualize ("ALL PEPFAR" as default or list OU names)}
}
\description{
Prepare viz data
}
\keyword{internal}
