% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_release.R
\name{publish_release}
\alias{publish_release}
\title{Publish Tidy UNAIDS Data as a GitHub Release}
\usage{
publish_release(df)
}
\arguments{
\item{df}{dataframe created by \code{munge_unaids}}
}
\description{
This is a developer's tool to take the output from EDMS that has been
tidied via \code{munge_unaids} and push it to GitHub as a
\href{https://github.com/USAID-OHA-SI/mindthegap/releases}{Release}. This is
performed once annually when the data are updated and then users will
access via \code{load_unaids}. This function posts versions for all countries
as well as PEPFAR only countries in csv and rds formats.
}
\examples{
 \dontrun{
  filepath <- "../DataList_10_1_2030-12_00_00-AM.csv"
  df <- munge_edms(filepath)
  publish_release(df)
}
}
\seealso{
\code{\link[=munge_unaids(), load_unaids]{munge_unaids(), load_unaids()}}
}
