% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unaids.R
\name{pull_estimates}
\alias{pull_estimates}
\title{Pull clean UNAIDS HIV Estimates}
\usage{
pull_estimates(pepfar_only = TRUE)
}
\arguments{
\item{pepfar_only}{filters dataset to only PEPFAR countries if TRUE (default = TRUE)}
}
\value{
df
}
\description{
Pull clean UNAIDS 2022 (1990-2022) estimates data. Wrapper around
\code{pull_unaids}.
}
\examples{
 \dontrun{
   pull_estimates()
}
}
