% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unaids.R
\name{pull_testtreat}
\alias{pull_testtreat}
\title{Pull clean UNAIDS HIV Test & Treat}
\usage{
pull_testtreat(pepfar_only = TRUE)
}
\arguments{
\item{pepfar_only}{filters dataset to only PEPFAR countries if TRUE (default = TRUE)}
}
\value{
df
}
\description{
Deprecated.
Pull clean UNAIDS 2022 (1990-2022) HIV Test & Treat data. Wrapper around
\code{pull_unaids}.
}
\examples{
 \dontrun{
   pull_testtreat()
}
}
