% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unaids.R
\name{pull_unaids}
\alias{pull_unaids}
\title{Pull clean UNAIDS 2022 Estimates}
\usage{
pull_unaids(data_type, pepfar_only = TRUE)
}
\arguments{
\item{data_type}{returns one of the 2 datasets available
eg "HIV Estimates", "HIV Test & Treat"}

\item{pepfar_only}{return dataset of only PEPFAR countries if TRUE (default = TRUE)}
}
\value{

}
\description{
Pull clean UNAIDS 2022 (1990-2021) estimates
}
\examples{
 \dontrun{
   pull_unaids(data_type = "HIV Estimates", pepfar_only = TRUE)
}

}
