% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{read_edms}
\alias{read_edms}
\title{Import EDMS Export}
\usage{
read_edms(path)
}
\arguments{
\item{path}{fileoath to the EDMS export (csv)}
}
\description{
Read in the data that has been downloaded from the
\href{https://edms.unaids.org/}{UNAIDS EDMS Database}
}
\keyword{internal}
