% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_rename}
\alias{read_rename}
\title{Import function for reading original UNAIDS estimates}
\usage{
read_rename(return_type)
}
\arguments{
\item{return_type}{Either "HIV Estimates" or "HIV Test & Treat"}
}
\value{
returns df to pass into munge_unaids()
}
\description{
Import function for reading original UNAIDS estimates
}
