% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{Title
Export a list of UN regions}
\format{
An object of class \code{character} of length 9.
}
\usage{
regions
}
\description{
This list is used to filter the UN regions
}
\keyword{datasets}
