% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{req_cols}
\alias{req_cols}
\title{EMDS Extract Required Columns}
\format{
Character vectors.
A vector of length 11
}
\usage{
req_cols
}
\description{
This is a list of the key columns that are needed through the munging process
for the clean export. The dataset originates from the
\href{https://edms.unaids.org/}{UNAIDS EDMS Database}.
}
\keyword{datasets}
