% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_info.R
\docType{data}
\name{source_note}
\alias{source_note}
\title{Title
UNAIDS Source Info}
\format{
An object of class \code{character} of length 1.
}
\usage{
source_note
}
\description{
Returns UNAIDS Source info for consistent sourcing notes
}
\keyword{datasets}
