% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{spread_values}
\alias{spread_values}
\title{Pivot Values
Spread estimates and bounds to their own columns. Create a flag where
estimate use < or >. Handle character truncation (eg "m").}
\usage{
spread_values(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Pivot Values
Spread estimates and bounds to their own columns. Create a flag where
estimate use < or >. Handle character truncation (eg "m").
}
\keyword{internal}
