% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{subset_cols}
\alias{subset_cols}
\title{Remove extraneous columns}
\usage{
subset_cols(df)
}
\arguments{
\item{df}{imported dataframe}
}
\description{
Keep only necessary columns that are required for output or are needed in the
process.
}
\keyword{internal}
