% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot.R
\name{subset_epi_data}
\alias{subset_epi_data}
\title{Subset data to select indicators needed}
\usage{
subset_epi_data(df)
}
\arguments{
\item{df}{UNAIDS data frame loaded from \code{load_unaids()}}
}
\description{
Filter down data to only what is needed for the epi control plots
}
\keyword{internal}
