% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plot.R
\name{tab_95s}
\alias{tab_95s}
\title{95's Table Plot}
\usage{
tab_95s(
  df,
  cntry,
  denom = c("PLHIV", "Relative"),
  grp = c("All", "Female 15+", "Male 15+", "Peds <15"),
  yr = NULL
)
}
\arguments{
\item{df}{dataframe from \code{load_unaids}}

\item{cntry}{PEPFAR country to visualize (list OU name)}

\item{denom}{which denominator/base to use, "PLHIV" (default) or "Relative"}

\item{grp}{age/sex group, c("All", "Female 15+", "Male 15+", "Peds <15")}

\item{yr}{year in question, defaults to max year}
}
\value{
gt table
}
\description{
This function creates a summary table showing OU progress toward the 95-95-95's
}
\examples{
 \dontrun{
   tab_95s(load_unaids(), "Kenya", denom = "PLHIV")
   tab_95s(load_unaids(), "Kenya", denom = "Relative")
}

}
\seealso{
\code{\link[=load_unaids]{load_unaids()}}
}
