% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{unaids_year}
\alias{unaids_year}
\title{Year UNAIDS data are released through}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
unaids_year
}
\description{
Year UNAIDS data are released through
}
\keyword{internal}
