% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_impatt.R
\name{update_impatt}
\alias{update_impatt}
\title{Update IMPATT Dataset}
\usage{
update_impatt(filepath, usaid_email)
}
\arguments{
\item{filepath}{filepath to IMPATT data (from pepfar-panorama.org)}

\item{usaid_email}{USAID email to store file on GDrive}
}
\description{
Update IMPATT Dataset
}
\examples{
 \dontrun{
   path <- "~/Data/MER_Structured_Datasets_NAT_SUBNAT_FY15-21_20200918_v2_1.rds"
   update_impatt(path, "twhitmore@usaid.gov") }
}
