% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_unaids.R
\name{update_unaids}
\alias{update_unaids}
\title{Update UNAIDS 90-90-90 data}
\usage{
update_unaids(filepaths, usaid_email)
}
\arguments{
\item{filepaths}{filepaths to all 3 UNAIDS files - ALL, Female, Male}

\item{usaid_email}{USAID email to store file on GDrive}
}
\description{
Update UNAIDS 90-90-90 data
}
\examples{
\dontrun{
   files <- list.files("data-raw", full.names = TRUE)
   update_unaids(files, "twhitmore@usaid.gov") }
}
