% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate Columns}
\usage{
validate_columns(df)
}
\arguments{
\item{df}{imported dataframe}
}
\description{
Ensure that all the columns that are needed exists in the import
}
\keyword{internal}
