% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_edms.R
\name{validate_countries}
\alias{validate_countries}
\title{Validate PEPFAR countries}
\usage{
validate_countries(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Ensure there are no missing PEPFAR countries from the EDMS pull
}
\keyword{internal}
