% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot.R
\name{validate_epi_disaggs}
\alias{validate_epi_disaggs}
\title{Validate Epi indicator/disaggs}
\usage{
validate_epi_disaggs(df)
}
\arguments{
\item{df}{UNAIDS data frame loaded from \code{load_unaids()}}
}
\description{
Validate Epi indicator/disaggs
}
\keyword{internal}
