% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_numeric_string.R
\name{validate_numeric_string}
\alias{validate_numeric_string}
\title{Validate Numberic String}
\usage{
validate_numeric_string(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Ensure that there are not extract character values that are accounted for and
handled during the munging process. Those values include: "<", ">", ".", " ",
"m", "k". To resolve, address these exceptions in convert_numeric_string()
and add them to the \code{expected_chars} list in validate_numeric_string().
}
\keyword{internal}
