% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{validate_path}
\alias{validate_path}
\title{Validate Path}
\usage{
validate_path(path)
}
\arguments{
\item{path}{filepath to the EDMS export (csv)}
}
\description{
Validate file path to EDMS output is a csv and exits.
}
\keyword{internal}
