% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{viz_impatt}
\alias{viz_impatt}
\title{Visualize share of PLHIV on ART}
\usage{
viz_impatt(ctry_sel)
}
\arguments{
\item{ctry_sel}{country to visualize}
}
\value{
gap plot
}
\description{
Visualize share of PLHIV on ART
}
\keyword{internal}
