% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{viz_unaids}
\alias{viz_unaids}
\title{Visualize 90-90-90 Trends from UNAIDS}
\usage{
viz_unaids(ctry_sel)
}
\arguments{
\item{ctry_sel}{country to visualize}
}
\value{
trend plots
}
\description{
Visualize 90-90-90 Trends from UNAIDS
}
\keyword{internal}
