% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot.R
\name{epi_plot}
\alias{epi_plot}
\title{Epidemic Control Plot}
\usage{
epi_plot(sel_cntry)
}
\arguments{
\item{sel_cntry}{country to visualize}
}
\value{
epidemic control plot showing trends in new infections and AIDS-related deaths
}
\description{
This function creates epidemic control curves for selected OU
}
\examples{
 \dontrun{
   epi_plot(sel_cntry = "Lesotho")
}

}
