% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot.R
\name{epi_plot}
\alias{epi_plot}
\title{Epidemic Control Plot}
\usage{
epi_plot(df, sel_cntry = c("All PEPFAR"))
}
\arguments{
\item{df}{UNAIDS data frame loaded from \code{load_unaids()}}

\item{sel_cntry}{PEPFAR country to visualize ("ALL PEPFAR" as default or list OU names)}
}
\value{
Epidemic control plot showing trends in new infections and total deaths to HIV population
}
\description{
This function creates epidemic control curves for "ALL PEPFAR" or selected OU's
}
\examples{
 \dontrun{
   df_unaids <- load_unaids(pepfar_only = FALSE)
   epi_plot(df_unaids)
   epi_plot(df_unaids, sel_cntry = "Lesotho")
   epi_plot(df_unaids, sel_cntry = c("South Africa", "Zambia", "Kenya", "Malawi"))
   epi_plot(df_unaids, sel_cntry = "USA") #breaks with non-PEPFAR countries
}

}
\seealso{
\code{\link[=load_unaids]{load_unaids()}}
}
