% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot.R
\name{epi_plot}
\alias{epi_plot}
\title{Epidemic Control Plot}
\usage{
epi_plot(sel_cntry = c("All PEPFAR"))
}
\arguments{
\item{sel_cntry}{PEPFAR country to visualize ("ALL PEPFAR" as default or list OU names)}

\item{df}{UNAIDS based data frame}
}
\value{
Epidemic control plot showing trends in new infections and total deaths to HIV population
}
\description{
This function creates epidemic control curves for "ALL PEPFAR" or selected OU's
}
\examples{
 \dontrun{
   epi_plot()
   epi_plot(sel_cntry = "Lesotho")
   epi_plot(sel_cntry = c("South Africa", "Zambia", "Kenya", "Malawi"))
   epi_plot(sel_cntry = "USA") #breaks with non-PEPFAR countries
}

}
