% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindthegap-package.R
\docType{package}
\name{mindthegap-package}
\alias{mindthegap}
\alias{mindthegap-package}
\title{mindthegap: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USAID-OHA-SI/mindthegap}
  \item Report bugs at \url{https://github.com/USAID-OHA-SI/mindthegap/issues}
}

}
\author{
\strong{Maintainer}: Aaron Chafetz \email{achafetz@usaid.ogv}

}
\keyword{internal}
