% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_unaids.R
\name{munge_unaids}
\alias{munge_unaids}
\title{Cleaning UNAIDS Data}
\usage{
munge_unaids(return_type, indicator_type)
}
\arguments{
\item{return_type}{Returns either 'HIV Estimates' or 'HIV Test & Treat' Data}

\item{indicator_type}{Returns either 'Integer' or 'Percent' indicator values}
}
\value{
df
}
\description{
(Updated July 2023)

This function fetches and cleans UNAIDS Estimates/Test and Treat Data
}
\examples{
 \dontrun{
   munge_unaids(return_type = "HIV Test & Treat", indicator_type = "Percent")
}

}
