% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unaids.R
\name{pull_unaids}
\alias{pull_unaids}
\title{Pull clean UNAIDS 2022 Estimates}
\usage{
pull_unaids(orginal_unaids = TRUE, data_type, pepfar_only = TRUE)
}
\arguments{
\item{orginal_unaids}{if TRUE, navigates to the original UNAIDS data received in July 2022.
If FALSE, navigates to the additional epicontrol, cascade, and prev indicators received in August 2022}

\item{data_type}{returns the available data set types
eg "HIV Estimates", "HIV Test & Treat" for original UNAIDS data (original_unaids = TRUE) or "epicontrol",
"cascade", "prev" for additional UNAIDS data (original_unaids = FALSE)}

\item{pepfar_only}{filters dataset to only PEPFAR countries if TRUE (default = TRUE)}
}
\value{

}
\description{
Pull clean UNAIDS 2022 (1990-2021) estimates
}
\examples{
 \dontrun{
   pull_unaids(data_type = "HIV Estimates", pepfar_only = TRUE)
}

}
