% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unaids.R
\name{pull_unaids}
\alias{pull_unaids}
\title{Pull clean UNAIDS 2021 Estimates for PEPFAR countries}
\usage{
pull_unaids(sheetname)
}
\arguments{
\item{sheetname}{returns the tab name from the Google Sheet
eg "HIV Estimates - Integer", "HIV Estimates - Percent", "Test & Treat - Integer", "Test & Treat - Percent"}
}
\value{

}
\description{
Pull clean UNAIDS 2021 (1990-2020) estimates for PEPFAR countries from Google Sheet
}
\examples{
 \dontrun{
   pull_unaids(sheetname = "HIV Estimates - Integer")
}

}
