% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unaids.R
\name{pull_unaids}
\alias{pull_unaids}
\title{Pull clean UNAIDS 2022 Estimates}
\usage{
pull_unaids(sheetname, pepfar_only = FALSE)
}
\arguments{
\item{pepfar_only}{return dataset of only PEPFAR countries if TRUE (default = FALSE)}

\item{data_type}{returns the data type requested
eg "HIV Estimates", "Test & Treat"}
}
\value{

}
\description{
Pull clean UNAIDS 2022 (1990-2021) estimates
}
\examples{
 \dontrun{
   pull_unaids(data_type = "HIV Estimates", pepfar_only = TRUE)
}

}
