% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unaids.R
\name{pull_unaids}
\alias{pull_unaids}
\title{Pull clean UNAIDS 2021 Estimates}
\usage{
pull_unaids(sheetname, pepfar_only = FALSE)
}
\arguments{
\item{sheetname}{returns the tab name from the Google Sheet
eg "HIV Estimates - Integer", "HIV Estimates - Percent", "Test & Treat - Integer", "Test & Treat - Percent"}

\item{pepfar_only}{return dataset of only PEPFAR countries if TRUE}
}
\value{

}
\description{
Pull clean UNAIDS 2021 (1990-2020) estimates
}
\examples{
 \dontrun{
   pull_unaids(sheetname = "HIV Estimates - Integer", pepfar_only = TRUE)
}

}
