% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_columns.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate Columns}
\usage{
validate_columns(df, expected_cols)
}
\arguments{
\item{df}{imported dataframe}

\item{expected_cols}{a string vector of the expected columns that should be
in the dataframe}
}
\description{
Ensure that all the columns that are needed exists in the import
}
\keyword{internal}
