% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_edms.R
\name{validate_ind_disaggs}
\alias{validate_ind_disaggs}
\title{Validate Indicator and Disaggregated included}
\usage{
validate_ind_disaggs(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Check if all expected indicators and their disaggregates are included and/or
any additional indicators/disaggregates are included
}
\keyword{internal}
